Game Engine Patch v1.1

+++ src/features/storySimulator/scenes/ObjectPlacement/ObjectPlacement.ts
@@ -4,9 +4,8 @@ import { AssetKey, AssetPath } from 'src/features/game/commons/CommonTypes';
 import GameInputManager from 'src/features/game/input/GameInputManager';
 import GameLayerManager from 'src/features/game/layer/GameLayerManager';
 import { Layer } from 'src/features/game/layer/GameLayerTypes';
-import GameSoundManager from 'src/features/game/sound/GameSoundManager';
+import SourceAcademyGame from 'src/features/game/SourceAcademyGame';
 import { mandatory } from 'src/features/game/utils/GameUtils';
-import { getStorySimulatorGame } from 'src/pages/academy/storySimulator/subcomponents/storySimulatorGame';
 
 import SSImageAssets from '../../assets/ImageAssets';
 import SSBackgroundManager from '../../background/SSBackgroundManager';
@@ -19,9 +18,12 @@ import { StorySimState } from '../../StorySimulatorTypes';
 import SSTransformManager from '../../transform/SSTransformManager';
 import objPlacementConstants from './ObjectPlacementConstants';
 
+/**
+ * Allow users to position objects, set backgrounds, and get
+ * the coordinates of the objects.
+ */
 export default class ObjectPlacement extends Phaser.Scene {
   public layerManager: GameLayerManager;
-  public soundManager: GameSoundManager;
   public inputManager: GameInputManager;
   private transformManager: SSTransformManager;
   private cursorModes: SSCursorMode | undefined;
@@ -39,7 +41,6 @@ export default class ObjectPlacement extends Phaser.Scene {
   constructor() {
     super('ObjectPlacement');
     this.layerManager = new GameLayerManager();
-    this.soundManager = new GameSoundManager();
     this.inputManager = new GameInputManager();
     this.objectManager = new SSObjectManager();
     this.bboxManager = new SSBBoxManager();
@@ -54,7 +55,6 @@ export default class ObjectPlacement extends Phaser.Scene {
 
   public init() {
     this.layerManager = new GameLayerManager();
-    this.soundManager = new GameSoundManager();
     this.inputManager = new GameInputManager();
     this.objectManager = new SSObjectManager();
     this.bboxManager = new SSBBoxManager();
@@ -68,8 +68,9 @@ export default class ObjectPlacement extends Phaser.Scene {
   }
 
   public create() {
+    SourceAcademyGame.getInstance().setCurrentSceneRef(this);
+
     this.layerManager.initialise(this);
-    this.soundManager.initialise(this, getStorySimulatorGame());
     this.inputManager.initialise(this);
     this.renderBackground();
     this.createUIButtons();
@@ -96,23 +97,16 @@ export default class ObjectPlacement extends Phaser.Scene {
 
   private createUIButtons() {
     const uiContainer = new Phaser.GameObjects.Container(this, 0, 0);
-    const backButton = new CommonBackButton(
-      this,
-      () => {
-        this.cleanUp();
-        getStorySimulatorGame().setStorySimState(StorySimState.Default);
-        this.scene.start('StorySimulatorMenu');
-      },
-      0,
-      0,
-      this.soundManager
-    );
+    const backButton = new CommonBackButton(this, () => {
+      this.cleanUp();
+      SourceAcademyGame.getInstance().setStorySimState(StorySimState.Default);
+      this.scene.start('StorySimulatorMenu');
+    });
 
     this.cursorModes = new SSCursorMode(
       this,
       objPlacementConstants.cursorModeXPos,
-      objPlacementConstants.cursorModeYPos,
-      this.soundManager
+      objPlacementConstants.cursorModeYPos
     );
     this.populateCursorModes();
 
@@ -189,6 +183,9 @@ export default class ObjectPlacement extends Phaser.Scene {
 
       // Erase Layers
       this.cursorModes.addCursorMode(this, SSImageAssets.eraseIcon.key, false, 'Erase all', () => {
+        const confirm = window.confirm('Are you sure you want to delete?');
+
+        if (!confirm) return;
         this.layerManager.clearSeveralLayers([Layer.Background, Layer.BBox, Layer.Objects]);
         this.objectManager.deleteAll();
         this.bboxManager.deleteAll();
@@ -200,7 +197,7 @@ export default class ObjectPlacement extends Phaser.Scene {
     }
   }
 
-  public getCursorManager = () => mandatory(this.cursorModes) as SSCursorMode;
+  public getCursorManager = () => mandatory(this.cursorModes);
 
   public getCoordinates(): number[] {
     return [this.input.x, this.input.y];
